// admin/src/app/@core/models/vw-relatorio-inscricoes.model.ts
export class VwRelatorioInscricoes {
    idUsuario: number;
    idInscricao: number;
    nome: string;
    telefoneCelular: string;
    telefoneFixo: string;
    curso: string;
    status: string;
    dataInscricao: Date;
    ciclo: string;
    ano: string;
    meioPagamento: string;
    meioPagamentoDesc: string;
    valorLiquido: string;
    valorBruto: string;
    qtdParcelas: string;
}